+++ Lesson25Servlet/src/main/java/main/HelloServlet.java
@@ -1,0 +1,35 @@
+package main;
+
+import java.io.*;
+
+import jakarta.servlet.http.*;
+import jakarta.servlet.annotation.*;
+
+// 1. Создать страницу логина с формой ввода имени пользователя и пароля.
+// 2. Создать сервлет для обработки формы логина и определения пользователя
+// (можно использовать предустановленные значения имени и пароля, в БД нет необходимости).
+//пункты 1 и 2 уже реализованы, можно взять за основу в моем репозитории!
+// 3. Создать защищенную страницу, попасть на которую можно только после логина.
+// 4. Все запросы к защищенной странице должны приводить к перенаправлению на страницу логина,
+// если пользователь еще не залогинен, а если залогинен – должны всегда быть успешными.
+@WebServlet(value = "/second-servlet")
+public class HelloServlet extends HttpServlet {
+    private String message;
+
+    public void init() {
+        message = "Hello Yauhen";
+    }
+
+    public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
+        response.setContentType("text/html");
+
+        // Hello
+        PrintWriter out = response.getWriter();
+        out.println("<html><body>");
+        out.println("<h1>" + message + "</h1>");
+        out.println("</body></html>");
+    }
+
+    public void destroy() {
+    }
+}
