+++ Lesson25Servlet/src/main/webapp/secondServlet.jsp
@@ -5,9 +5,9 @@
     <title>JSP - Hello World</title>
 </head>
 <body>
-<h1><%= "Hello World!" %>
+<h1><%= "Hello Developer" %>
 </h1>
 <br/>
-<a href="hello-servlet">Hello Servlet</a>
+<a href="second-servlet">Hello Servlet</a>
 </body>
 </html>
