+++ Lesson9/src/main/java/flowers/FlowerType.java
@@ -8,9 +8,12 @@ import lombok.ToString;
 @Getter
 @ToString
 public enum FlowerType {
-    ROSE(15, "ROSE"), LILY(7, "LILY"),
-    ASTER(5, "ASTER"), GERBERA(5, "GERBERA"),
-    TULIP(8, "TULIP"), CARNATION(11, "CARNATION");
+    ROSE(15, "ROSE"),
+    LILY(7, "LILY"),
+    ASTER(5, "ASTER"),
+    GERBERA(5, "GERBERA"),
+    TULIP(8, "TULIP"),
+    CARNATION(11, "CARNATION");
     private final int price;
     private final String name;
 }
